+++ karbide-core/src/commonMain/kotlin/dev/karmakrafts/karbide/BitSourceLeExtensions.kt
@@ -1,5 +1,73 @@
 package dev.karmakrafts.karbide
 
+// ------------------------------ Peeking ------------------------------
+
+/**
+ * Read a [Short] (16 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The short read.
+ */
+fun BitSource.peekShortLe(): Short = peekBits(Short.SIZE_BITS).toShort().reverseBytes()
+
+/**
+ * Read an [Int] (32 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The integer read.
+ */
+fun BitSource.peekIntLe(): Int = peekBits(Int.SIZE_BITS).toInt().reverseBytes()
+
+/**
+ * Read a [Long] (64 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The long read.
+ */
+fun BitSource.peekLongLe(): Long = peekBits(Long.SIZE_BITS).toLong().reverseBytes()
+
+/**
+ * Read a [Float] (32 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The float read.
+ */
+fun BitSource.peekFloatLe(): Float = Float.fromBits(peekBits(Float.SIZE_BITS).toInt().reverseBytes())
+
+/**
+ * Read a [Double] (64 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The double read.
+ */
+fun BitSource.peekDoubleLe(): Double = Double.fromBits(peekBits(Double.SIZE_BITS).toLong().reverseBytes())
+
+/**
+ * Read a [UShort] (16 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The unsigned short read.
+ */
+fun BitSource.peekUShortLe(): UShort = peekBits(UShort.SIZE_BITS).toUShort().reverseBytes()
+
+/**
+ * Read a [UInt] (32 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The unsigned integer read.
+ */
+fun BitSource.peekUIntLe(): UInt = peekBits(UInt.SIZE_BITS).toUInt().reverseBytes()
+
+/**
+ * Read a [ULong] (64 bits) from the [BitSource] in little endian byte order
+ * without consuming the bits.
+ *
+ * @return The unsigned long read.
+ */
+fun BitSource.peekULongLe(): ULong = peekBits(ULong.SIZE_BITS).reverseBytes()
+
+// ------------------------------ Reading ------------------------------
+
 /**
  * Read a [Short] (16 bits) from the [BitSource] in little endian byte order.
  *